# pkg/trace: add appsec span tag obfuscation