--- a/cmd/trace-agent/config/config.go
+++ b/cmd/trace-agent/config/config.go
@@ -20,6 +20,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/cmd/trace-agent/internal/osutil"
 	coreconfig "github.com/DataDog/datadog-agent/pkg/config"
+	"github.com/DataDog/datadog-agent/pkg/obfuscate"
 	"github.com/DataDog/datadog-agent/pkg/otlp"
 	"github.com/DataDog/datadog-agent/pkg/proto/pbgo"
 	"github.com/DataDog/datadog-agent/pkg/tagger"
@@ -290,12 +291,10 @@ func applyDatadogConfig(c *config.AgentConfig) error {
 			c.Obfuscation.CreditCards.Luhn = coreconfig.Datadog.GetBool("apm_config.obfuscation.credit_cards.luhn")
 		}
 	}
-
 	c.Obfuscation.AppSec = config.AppSecObfuscationConfig{
-		ParameterKeyRegexp:   compileAppSecRegexp(coreconfig.Datadog.GetString(coreconfig.AppSecConfigKeyObfuscationParameterKeyRegexp)),
-		ParameterValueRegexp: compileAppSecRegexp(coreconfig.Datadog.GetString(coreconfig.AppSecConfigKeyObfuscationParameterValueRegexp)),
+		ParameterKeyRegexp:   obfuscate.CompileRegexp(coreconfig.Datadog.GetString("appsec_config.obfuscation.parameter_key_regexp")),
+		ParameterValueRegexp: obfuscate.CompileRegexp(coreconfig.Datadog.GetString("appsec_config.obfuscation.parameter_value_regexp")),
 	}
-
 	if coreconfig.Datadog.IsSet("apm_config.filter_tags.require") {
 		tags := coreconfig.Datadog.GetStringSlice("apm_config.filter_tags.require")
 		for _, tag := range tags {
@@ -601,19 +600,4 @@ func acquireHostnameFallback(c *config.AgentConfig) error {
 	}
 	log.Debugf("Acquired hostname from core agent (%s): %q.", c.DDAgentBin, c.Hostname)
 	return nil
-}
-
-// compileAppSecRegexp compiles the given expression into a Go regular expression. It returns nil when the given
-// expression cannot be compiled and logs the compilation error. The empty string expression is considered the special
-// expression value to use in order to disable the regular expression.
-func compileAppSecRegexp(expr string) *regexp.Regexp {
-	if expr == "" {
-		return nil
-	}
-	re, err := regexp.Compile(expr)
-	if err != nil {
-		log.Errorf("Failed to parse the regular expression `%s`: %s", expr, err)
-		return nil
-	}
-	return re
 }