… cmd/trace-agent/config/testdata/full.yaml
… pkg/flare/envvars.go
… pkg/obfuscate/appsec_test.go
… pkg/obfuscate/json_scanner.go
… pkg/trace/agent/obfuscate.go
… pkg/trace/agent/obfuscate_test.go
… pkg/trace/config/config.go
… pkg/trace/config/config_test.go
… releasenotes/notes/trace-agent-appsec-events-obfuscation-a4c1705aa766475c.yaml