<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Draft PRs should be prefixed with `[WIP]` in their title.

-->
### What does this PR do?

Obfuscate `_dd.appsec.json` tags against sensitive data.
AppSec events contain security rule inputs, called parameters here, that may contain sensitive data which need to be obfuscated. It is made of two regular expressions: one for the parameter keys (ie. original security rule parameter field names) and another one for the parameter value and highlight values.
The obfuscator is always enabled and obfuscates values whose keys can be `password`, `pwd`, `token`, `api_token` and so on.

**Important note:** the implementation is permissive enough to let new and different appsec json payloads to happen without having to immediately update the agent.

The obfuscation algorithm uses the existing json scanner and scans each parameter of each appsec security events. It searches for all of the appsec rule parameters the payload may contain so that:
1. It firstly tries to match the parameter key names against the key regular expression. When a key matches, the full parameter value and highlights are obfuscated regardless of the value regular expression.
2. If no parameter key was matched, then the value regular expression is executed on the parameter value and highlights. 

### Benchmarks

#### Unquote implementation comparisons

The performance of this obfuscator is driven by the function `unquote()` which is called everywhere in order to unquote the JSON string values, compare them and ultimately obfuscate them.
Therefore, the following benchmark results compares what happens when using `encoding/json`'s internal `unquote()` function (new) versus calling `encoding/json.Unmarshal()` on a Go string to avoid copying this internal code (old): 

```
name                       old time/op    new time/op    delta
Obfuscator/matching-4        76.5µs ± 0%    78.7µs ± 0%   +2.85%
Obfuscator/not-matching-4    69.5µs ± 0%    50.2µs ± 0%  -27.73%

name                       old alloc/op   new alloc/op   delta
Obfuscator/matching-4        25.4kB ± 0%    32.6kB ± 0%  +28.34%
Obfuscator/not-matching-4    23.6kB ± 0%    10.7kB ± 0%  -54.59%

name                       old allocs/op  new allocs/op  delta
Obfuscator/matching-4           354 ± 0%       294 ± 0%  -16.95%
Obfuscator/not-matching-4       354 ± 0%       138 ± 0%  -61.02%
```

The resulting performance boost mainly comes from the fact we avoid the string <-> []byte copies involved by the call to `json.Unmarshal()`, whereas the compiler is able to optimize and avoid those when `encoding/json`'s `unquote` function is copied into the obfuscator package (likely thanks to inlining + escape analysis + immutability checks).

This is the reason why this PR is opened with the option of copying and using `encoding/json`'s `unquote()` function.


#### Obfuscator implementation comparisons

The following results compare the json-scanner-based implementation (using the previous unquote optimization) to the basic implementation decoding the full AppSec JSON payload in order to obfuscate it:

```
name                       old time/op    new time/op    delta
Obfuscator/matching-4         114µs ± 0%      79µs ± 0%  -31.10%
Obfuscator/not-matching-4    63.5µs ± 0%    50.2µs ± 0%  -20.92%

name                       old alloc/op   new alloc/op   delta
Obfuscator/matching-4        41.7kB ± 0%    32.6kB ± 0%  -21.73%
Obfuscator/not-matching-4    23.4kB ± 0%    10.7kB ± 0%  -54.13%

name                       old allocs/op  new allocs/op  delta
Obfuscator/matching-4           786 ± 0%       294 ± 0%  -62.60%
Obfuscator/not-matching-4       466 ± 0%       138 ± 0%  -70.39%
```

This confirms it is a good technical choice in the end for performances, despite the much (much) harder implementation.

### Motivation

Sensitive data obfuscation.

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->
The longer term better solution for this is to move to the new `meta_struct` span tags which would allow to walking a Go value instead of a bare JSON payload that we need to scan here.

### Describe how to test/QA your changes

- This change is 100% covered by the Go tests.
- System tests were added to check the sensitive data gets removed: https://github.com/DataDog/system-tests/pull/186

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [x] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [x] The appropriate `team/..` label has been applied, if known.
- [x] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [x] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [x] Changed code has automated tests for its functionality.
- [x] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [x] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [x] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.