@Julio-Guerra 2022-03-29
Closing the PR after finding out that we could do it at the AppSec WAF level when it generates the AppSec events. This solution avoids all the drawbacks of doing it offline, with the implications it has on the sensitive data lookup.